Added already installed checks

+++ app-installs/vscode.sh
@@ -2,6 +2,11 @@
 
 set -e
 
-wget -qO vscode.deb 'https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64'
-sudo apt install -y ./vscode.deb
-rm -f vscode.deb
+CNT=$(apt-cache pkgnames "code" | grep -e '^code$')
+if [ $? -ne 0 ]; then
+    wget -qO vscode.deb 'https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64'
+    sudo apt install -y ./vscode.deb
+    rm -f vscode.deb
+else
+    echo VSCode already installed
+fi
